// types.ts
export enum OrderStatus {
  PENDING = 'PENDING',
  RECEIVED = 'RECEIVED',
  PREPARING = 'PREPARING',
  DELIVERING = 'DELIVERING',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED'
}

export const OrderStatusLabels = {
  [OrderStatus.PENDING]: 'Aguardando Confirmação',
  [OrderStatus.RECEIVED]: 'Recebido',
  [OrderStatus.PREPARING]: 'Em Preparo',
  [OrderStatus.DELIVERING]: 'Em Entrega',
  [OrderStatus.COMPLETED]: 'Concluído',
  [OrderStatus.CANCELLED]: 'Cancelado'
};

export enum PaymentMethod {
  CASH = 'CASH',
  CARD = 'CARD',
  PIX = 'PIX',
  WALLET = 'WALLET',
  OTHER = 'OTHER'
}

export const PaymentMethodLabels = {
  [PaymentMethod.CASH]: 'Dinheiro',
  [PaymentMethod.CARD]: 'Cartão',
  [PaymentMethod.PIX]: 'Pix',
  [PaymentMethod.WALLET]: 'Carteira Digital',
  [PaymentMethod.OTHER]: 'Outro'
};

export enum ProductCategory {
  SNACK = 'Lanches',
  PIZZA = 'Pizzas',
  DRINK = 'Bebidas',
  DESSERT = 'Sobremesas'
}

export type OrderSource = 'STORE' | 'IFOOD' | 'WHATSAPP' | 'DIGITAL_MENU' | 'WHATSAPP_BOT';
// ...
export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  category: ProductCategory;
  imageUrl: string;
  stock: number;
  ingredients: string[];
}

export interface OrderItem {
  productId: string;
  productName: string;
  quantity: number;
  unitPrice: number;
  notes?: string;
  // For pizzas
  isHalfHalf?: boolean;
  secondFlavorId?: string;
  secondFlavorName?: string;
  extras?: string[]; // e.g., Stuffed Crust
}
// ...
export interface Order {
  id: string;
  displayId?: number; // Sequential ID (Backend)
  source: OrderSource;
  clientId: string;
  clientName: string;
  clientPhone: string;
  deliveryAddress: string;
  deliveryAddressReference?: string; // Saved reference point
  items: OrderItem[];
  subtotal: number;

  deliveryFee: number; // What the CLIENT pays
  driverFee?: number;  // What the STORE pays the driver (Internal cost)
  driverPaid?: boolean; // NEW: If the driver has been paid for this order

  discount: number;
  total: number;
  status: OrderStatus;
  paymentMethod: PaymentMethod;
  changeFor?: number; // Troco para quanto (if cash)
  paymentStatus: 'Pending' | 'Paid';

  // NEW: Driver Link
  driverId?: string;
  driverName?: string;

  createdAt: string; // ISO date
  updatedAt: string;
}

export interface DeliveryRange {
  id: string;
  minKm: number;
  maxKm: number;
  price: number;
}

export interface StoreSettings {
  name: string;
  address: string;
  logoUrl?: string;
  managerPassword?: string;
  contactPhone?: string; // WhatsApp number
  isStoreOpen?: boolean; // Status da Loja (Aberta/Fechada)

  // WhatsApp Connection Status
  whatsappStatus?: 'CONNECTED' | 'DISCONNECTED' | 'PAIRING';

  deliveryRanges?: DeliveryRange[]; // Fee charged to CUSTOMER
  driverFeeRanges?: DeliveryRange[]; // Fee paid to DRIVER

  integrations?: {
    ifoodEnabled: boolean;
    whatsappEnabled: boolean;
  };

  latitude?: number;
  longitude?: number;
}